Refactor RunShortLinkRepTest
